+++ app/routes.js
@@ -1,4 +1,4 @@
 const users = require('./controllers/users.js'),
-  signUpValidator = require('./middlewares/signUpValidator').handle;
+  signUpValidator = require('./middlewares/validator').signUpValidator;
 
 exports.init = app => app.post('/users', [signUpValidator], users.create);
