+++ app/services/users.js
@@ -1,15 +1,12 @@
 const user = require('../models').user,
   logger = require('../logger'),
-  errors = require('../errors'),
-  bcryptService = require('../services/bcrypt');
+  errors = require('../errors');
 
 exports.create = userFields =>
-  user
-    .create({ ...userFields, password: bcryptService.encryptPassword(userFields.password) })
-    .catch(error => {
-      logger.info(`Failed to create the user. ${error}`);
-      throw errors.databaseError(error.message);
-    });
+  user.create({ ...userFields }).catch(error => {
+    logger.info(`Failed to create the user. ${error}`);
+    throw errors.databaseError(error.message);
+  });
 
 exports.findBy = condition =>
   user.findOne({ where: condition }).catch(error => {
